Completed Assignment
Properly used the connect function and the variable mapStateToProps to assign state through the combined reducer by referring to state as a property of the charsReducer(which is where the reducer code exists which returns that specific state). This is done in this example as characters: state.charReducer.characters in the mapStateToProps variable.